Fix for multiple applications of the map per turn.
Store lastTurnNumber in the OTM.  When it's applied: increment it
internally.  If TurnsTaken() - 1 == lastTurnNumber (-1 because in the
terminator, which comes before the teleporter, the turn number is
incremented), then it has not yet been applied.  It is applied,
lastTurnNumber now equals TurnsTaken(), and instead simply the cached
values are returned.