test for userfile s and update the test lattice with a massive bend
this is necessary because i want to be able to see the difference
between z and s, to check that s has been implemented correctly, and
the difference is only clear to the eye when there is a large angle.
when the distrFileFormat is simply "z[m]" then the beam starts way
outside of the machine, missing the bend.  when it is "s[m]" then the
beam correctly starts in the middle of the dipole bend.