length safety to placement of BLMs and recentering for lhc magnets
lhcleft and lhcright for example are offset.  so if you were to place
the BLM without an xOffset then it would be directly above the
beampipe, i.e. offset transversely w.r.t the magnet geometry.  instead
it is preferable to have the BLM centred, not directly above the beampipe.