update blm attachment test so bdsi has to look forwards & backwards.
this is the more intricate and general case and should be tested in
favour of the previous one, where a blm was just attached to a single
quadurpole.  now we attach it to a short drift with large quadrupoles
either side of it such that there would definitely be overlaps if not
for looking forward and backward along the beamline when getting the
transverse placement.