Implementing warning messages in case one is using time modulation on a field map that varies in time
The BDSArrayCoords classes are extended by functions giving back the dimension (x, y, z, t) of the
variable they are describing. This is used in order to get the dimension of a given field map.
In case the field map is already a function of t and one is trying to use a time modulation,
BDSIM throws a warning message and the additinal time variation is omitted.