Removing the GetDimension-Functions and making use of inheritance from 4D-class
The single GetDimension-Functions in the BDSArrayCoords1D-3D classes are removed.
Instead, the inherited versions from BDSArrayCoords4D are used to check if one
of the dimensions of a field map is time. This is used in order to avoid
modulating a field map with time when it's already a function of t.