Exception if it is requested to skip/ignore more lines than we have in the file
Now, we throw an exception if it is requested to skip and/or ignore more lines
than there are in the userfile. To show the behaviour when doing this, there
is the example userfile-bad-skipping (included as simple_fail in the testing),
and the right usage is shown in the userfile-skip-lines.