Detection of units of columns
In case, we have a field map that depends on t, we need to mulitply
the entries by CLHEP::s and not CLHEP::cm. For all kinds of field
maps, 1D to 4D, a check is included, to determine the dimension type
of the first and last column as by construction t can only be the
first or last columns (columns have to be in ascending or descending
order).